Task5/requirements.txt
azure-cosmos==5.3.0
